requirements.txt
flask==3.0.3
werkzeug==3.0.4
tensorflow==2.17.0
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3